moved karaoke models to different folder
