+++ src/models/karaoke.py
